refactor: 알림 저장과 실시간 알림 서비스를 로직을 Event-Driven 구조로 리팩토링
(#48)
